fix: wire up seed data on startup and fix bcrypt/passlib crash
- init_db.py imported from app.api.v1.services/schemas, which no
  longer exist after the domain-driven refactor — prestart.sh would
  have crashed on every fresh container start. Fixed import paths and
  wired init_seed_data() in (idempotent: skips if any entity exists).
- app/core/security.py had no compatibility shim for bcrypt >= 4.1
  (dropped __about__.__version__, changed 72-byte truncation
  behavior) — passlib's own internal self-test crashes on first use
  in any fresh environment. The exact same shim already existed in
  tests/conftest.py (tests never hit this because of it) but was
  never applied to the actual running app. Moved it to security.py
  and removed the now-redundant copy from conftest.py.

Verified against a fresh Postgres 15 + clean pip install: migrations
apply, superuser + 11 entities seed correctly, idempotent on a
second run (no duplicates), full test suite passes (145/145, 82.5%
coverage).
