+++ app/db/init_db.py
@@ -1,6 +1,9 @@
 from sqlalchemy.orm import Session
 
-from app.api.v1 import services, schemas
+from app.api.v1.domains.users.services.user import user as user_service
+from app.api.v1.domains.users.schemas.user import UserCreate
+from app.api.v1.entities.models.entity import Entity
+from app.api.v1.entities.seed import init_seed_data
 from app.core.config import settings
 from app.db import base  # noqa: F401
 
@@ -15,11 +18,16 @@ def init_db(db: Session) -> None:
     # the tables un-commenting the next line
     # Base.metadata.create_all(bind=engine)
 
-    user = services.user.get_by_email(db, email=settings.FIRST_SUPERUSER)
+    user = user_service.get_by_email(db, email=settings.FIRST_SUPERUSER)
     if not user:
-        user_in = schemas.UserCreate(
+        user_in = UserCreate(
             email=settings.FIRST_SUPERUSER,
             password=settings.FIRST_SUPERUSER_PASSWORD,
             is_superuser=True,
         )
-        user = services.user.create(db, obj_in=user_in)  # noqa: F841
+        user = user_service.create(db, obj_in=user_in)  # noqa: F841
+
+    # Seed the mythology catalog on first run only (idempotent: skip if any
+    # entity already exists, since init_seed_data() doesn't check for dupes)
+    if db.query(Entity).first() is None:
+        init_seed_data(db)
